Refactorizado método eliminar documentación

+++ APIRest/Controllers/DocumentacionesController.cs
@@ -99,24 +99,35 @@ namespace APIRest.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<JsonResult> Eliminar(int id)
+        public async Task<ActionResult> Eliminar(int id)
         {
-            Documentacion documentacion = await _contexto.Documentaciones
-                                                         .FirstOrDefaultAsync(documentacion => documentacion.Id == id);
+            Archivo documentacion = await _repositorioDocumentaciones.ObtenerPorId(id);
+
+            if (documentacion is null)
+                return NotFound($"No existe la documentación con id {id}");
+
+            if (System.IO.File.Exists(documentacion.UrlArchivo))
+            {
+                try
+                {
+                    System.IO.File.Delete(documentacion.UrlArchivo);
+                }
+                catch (Exception)
+                {
+                    return StatusCode(500, "Ha habido un error al eliminar el archivo");
+                }                
+            }
+
             try
             {
-                if (System.IO.File.Exists(documentacion.UrlArchivo))                
-                    System.IO.File.Delete(documentacion.UrlArchivo);                
-                
-                _contexto.Remove(documentacion);
-
-                await _contexto.SaveChangesAsync();
-                return new JsonResult(true);
+                await _repositorioDocumentaciones.Eliminar(documentacion);
             }
-            catch (Exception ex)
+            catch (Exception)
             {
-                return new JsonResult(false);
+                return StatusCode(500, "Ha habido un error al eliminar la documentación");
             }
+
+            return Ok(true);
         }
     }
 }
